move preview command to Descriptions/main.py
