+++ Descriptions/preview.py
@@ -1,19 +1,5 @@
-import argparse
-import json
-import sys
 import openpyxl
 from openpyxl.worksheet.formula import ArrayFormula
-from pathlib import Path
-
-
-def create_argument_parser():
-    parser = argparse.ArgumentParser()
-    subparsers = parser.add_subparsers()
-    parser_preview = subparsers.add_parser("preview", help="Create xlsx preview")
-    parser_preview.add_argument("in_file", help="Input json file")
-    parser_preview.add_argument("out_file", help="Output xlsx file")
-    parser_preview.set_defaults(func=create_preview_xlsx)
-    return parser
 
 
 def get_original_formula(row_index):
@@ -29,20 +15,6 @@ def get_translation_formula(ref_sheet_name, row_index):
         f"D{row_index}",
         f'=TEXTJOIN("", FALSE, IFERROR({search}, {desc_cells}))',
     )
-
-
-def create_preview_xlsx(args):
-    in_filename = args["in_file"]
-    out_filename = args["out_file"]
-    file_stem = Path(in_filename).stem
-
-    in_file = open(in_filename, encoding="utf8")
-    json_object = json.load(in_file)
-    in_file.close()
-
-    workbook = openpyxl.Workbook()
-    create_preview_worksheet(workbook, file_stem, json_object["data"])
-    workbook.save(out_filename)
 
 
 def create_preview_worksheet(workbook, sheet_name, data):
@@ -91,13 +63,3 @@ def format_preview_worksheet(worksheet):
     for row in worksheet["C:D"]:
         for cell in row:
             cell.alignment = alignment
-
-
-def main():
-    parser = create_argument_parser()
-    args = parser.parse_args(sys.argv[1:])
-    args.func(vars(args))
-
-
-if __name__ == "__main__":
-    main()
